+++ step_leapfrog.c
@@ -5,10 +5,10 @@
 #include <stdio.h>
 #include <math.h>
 void xOne(int particles, struct arrays *particleData){
-     float partx;
-     float partv;
-     float parta;
-     float x1;
+     double partx;
+     double partv;
+     double parta;
+     double x1;
      // for all particles 
      for (int i=0; i<particles; i++){
          partx = particleData->x[i];
@@ -16,17 +16,17 @@ void xOne(int particles, struct arrays *particleData){
          parta = particleData->a[i];
          x1 = partx + dt*partv + 0.5*(dt*dt)*parta;
          // handle periodic boundary
-         float deltaf = fabs(x1-xmax);
+         double deltaf = fabs(x1-xmax);
          if (x1 > xmax){
              printf("Moved Max! \n");
              printf("x1: %e \n",x1);
              printf("xmax: %e \n", xmax);
-             float deltax = x1-xmax;
+             double deltax = x1-xmax;
              printf("deltax: %e  \n",deltax);
              x1 = xmin + deltax;
          } else if (x1 < xmin){
              printf("Moved Min! \n");
-             float deltax = fabs(x1-xmin);
+             double deltax = fabs(x1-xmin);
              x1 = xmax - deltax;
          }
          particleData->x[i] = x1;
@@ -36,9 +36,9 @@ void xOne(int particles, struct arrays *particleData){
 }
 
 void vStar(int particles, struct arrays *particleData){
-    float partv;
-    float vstar;
-    float parta;
+    double partv;
+    double vstar;
+    double parta;
 
     for (int i=0; i<particles; i++){
         partv = particleData->v[i];
@@ -49,10 +49,10 @@ void vStar(int particles, struct arrays *particleData){
 
 }
 // copying array is a little sloppy 
-void vOne(int particles, struct arrays *particleData, float oldA[]){
-    float vstar;
-    float v1;
-    float parta;
+void vOne(int particles, struct arrays *particleData, double oldA[]){
+    double vstar;
+    double v1;
+    double parta;
 
     for (int i=0; i<particles; i++){
         vstar = particleData->v[i];
@@ -68,7 +68,7 @@ void step(int particles, struct arrays *particleData){
     vStar(particles,particleData);
     
     // store old a
-    float a0[particles];
+    double a0[particles];
     for (int i=0; i<particles; i++){
         a0[i] = particleData->a[i];
     }
